Move the api-fetch untility into the sync client
