+++ src/client/sync-client.ts
@@ -1,15 +1,21 @@
+import { asMaybe, Cleaner } from 'cleaners'
+import crossFetch from 'cross-fetch'
+
 import { EdgeServers } from '../types/base-types'
 import {
   asGetStoreResponse,
   asPostStoreResponse,
   asPutStoreResponse,
+  asServerErrorResponse,
+  GetStoreParams,
   GetStoreResponse,
   PostStoreBody,
+  PostStoreParams,
   PostStoreResponse,
+  PutStoreParams,
   PutStoreResponse
 } from '../types/rest-types'
-import { apiRequest } from '../util/api-request'
-import { CommonOptions } from '../util/common'
+import { CommonOptions, noOp } from '../util/common'
 import { syncKeyToRepoId } from '../util/security'
 import { shuffle } from '../util/shuffle'
 import { makeInfoClient } from './info-client'
@@ -118,4 +124,59 @@ export function makeSyncClient(opts: SyncClientOptions = {}): SyncClient {
       throw error
     }
   }
+}
+
+type ApiRequestBody = PostStoreBody
+type ApiRequestParams = GetStoreParams | PostStoreParams | PutStoreParams
+
+interface ApiRequest {
+  method: string
+  url: string
+  numbUrl?: string // Clean URL for logging
+  body?: ApiRequestBody
+  params?: ApiRequestParams
+  headers?: { [key: string]: string }
+}
+
+async function apiRequest<ApiResponse>(
+  request: ApiRequest,
+  asApiResponse: Cleaner<ApiResponse>,
+  opts: CommonOptions = {}
+): Promise<ApiResponse> {
+  const { log = noOp, fetch = crossFetch } = opts
+  const { method, url, body, numbUrl = url, headers = {} } = request
+
+  const start = Date.now()
+  const response = await fetch(url, {
+    method,
+    headers: {
+      'Content-Type': 'application/json',
+      ...headers
+    },
+    body: JSON.stringify(body)
+  })
+  const timeElapsed = Date.now() - start
+
+  log(`${method} ${numbUrl} returned ${response.status} in ${timeElapsed}ms`)
+
+  const responseBody = await response.text()
+
+  if (!response.ok)
+    throw new Error(
+      `Failed request ${method} ${numbUrl} failed ${response.status}: ${responseBody}`
+    )
+
+  const errorResponse = asMaybe(asServerErrorResponse)(responseBody)
+
+  if (errorResponse != null) {
+    throw new Error(
+      `Failed request ${method} ${numbUrl} failed ${response.status}: ${errorResponse.message}`
+    )
+  }
+
+  const responseData = asApiResponse(
+    responseBody.trim() !== '' ? JSON.parse(responseBody) : undefined
+  )
+
+  return responseData
 }
